dev_requirements.txt
appdirs==1.4.4
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
dateparser==0.7.4
idna==2.9
importlib-metadata==1.6.0
Jinja2==2.11.2
lxml==4.5.1
MarkupSafe==1.1.1
more-itertools==8.3.0
mutagen==1.44.0
packaging==16.8
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
pytest-cov==2.9.0
pytest-mock==3.1.0
python-dateutil==2.6.0
python-slugify==4.0.0
pytz==2017.2
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
requests-mock==1.8.0
six==1.10.0
text-unidecode==1.3
tzlocal==2.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
